+++ src/app/routes/private/routes/integrantes.private.js
@@ -3,5 +3,6 @@ const route = express.Router()
 const IntegrantesController = require('../../../controllers/integrantes')
 
 route.post('/criar', IntegrantesController.create)
+route.get('/listar-todos', IntegrantesController.getWithParams)
 
 module.exports = route