+++ CHANGELOG.md
@@ -4,3 +4,4 @@ Log de alterações da aplicação
 
 ### Versão 1.0
 * **Contatos:** Implementado CRUD completo e separado em rotas públicas e privadas.
+* **Sobre:** Desenvolvido o metodo create no módulo sobre.