+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # GPES API
 
 Log de alterações da aplicação
+* **Publicações:** Implementado o método Create