+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # GPES API
 
 Log de alterações da aplicação
-* **Publicações:** Implementado o método Create
+* **Publicações:** Implementado o método Create.
+* **Categorias:** Reaproveitando a implementação de categories da AsiloWebAPI para implementar os métodos de criar e listar as categorias.