+++ CHANGELOG.md
@@ -5,3 +5,5 @@ Log de alterações da aplicação
 ### Versão 1.0
 * **Contatos:** Implementado CRUD completo e separado em rotas públicas e privadas.
 * **Processo Seletivo:** Criado modelo, controller com os métodos de listar todos e criar, validador de título único, criação do arquivo de rotas para o módulo e alteração no arquivo de rotas privadas para inserir a rota de Processo Seletivo.
+* **Integrantes:** Implementando create e getall nas rotas privadas.
+