+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # GPES API
 
 Log de alterações da aplicação
-* **Publicações:** Implementado o método Create.
-* **Categorias:** Reaproveitando a implementação de categories da AsiloWebAPI para implementar os métodos de criar e listar as categorias.
+* **Publicações:** Implementado o método Create e getWithParams para listar todos, validador de título único.
+* **Categorias:** Reaproveitando a implementação de categories da AsiloWebAPI para implementar os métodos de criar e listar as categorias, validador de nome único.